+++ core/4_Callback/Hook.lua
@@ -31,8 +31,9 @@ run_once(function()
 end)
 
 local banned_scripts = {
-    gm.constants.step_actor,    -- Bad for performance
-    gm.constants.draw_actor,    -- Bad for performance
+    gm.constants.step_actor,        -- Bad for performance
+    gm.constants.draw_actor,        -- Bad for performance
+    gm.constants.actor_heal_raw,    -- Potentially bad for performance
 }