+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### v0.1.33
 - Actor : Fix `fire_*` methods not accepting Sprite wrappers
+- Hook : Add `actor_heal_raw` to banlist for potential lag
 
 ### v0.1.32
 - Actor : Fix `heal` crashing for host in multiplayer